feat(monitoring): integration with siteup
