+++ src/constants/featureFlags.ts
@@ -4,4 +4,5 @@ export const FEATURE_FLAGS = {
   IS_SHOW_STAGING_BUILD_STATUS_ENABLED: "is_show_staging_build_status_enabled",
   IS_CLOUDMERSIVE_ENABLED: "is_cloudmersive_enabled",
   IS_LOCAL_DIFF_ENABLED: "is_local_diff_enabled",
+  IS_MONITORING_ENABLED: "is_monitoring_enabled",
 } as const
