+++ src/types/featureFlags.ts
@@ -13,6 +13,7 @@ export interface FeatureFlags {
   is_show_staging_build_status_enabled: boolean
   is_cloudmersive_enabled: CloudmersiveConfigType
   is_local_diff_enabled: boolean
+  is_monitoring_enabled: boolean
 }
 
 // List of attributes we set in GrowthBook Instance in auth middleware
